pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "google-genai>=1.56.0",
    "libsql>=0.1.11",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "ruff>=0.13.2",
    "ty>=0.0.1a21",
]
